Code now in a compilable, executable state for Win32 and Linux/Cygwin/MinGW.
I've been debugging the round execution code, and there is definitely something messed up with the pointers somewhere. The code can correctly find and execute an ability deemed to be the best (... 1>1?), but it doesn't seem to properly affect the correct character. I'm guessing it's hitting the incorrect Character pointers.

+++ engine/BattleEngine/Source/System/RGR_List.cpp
@@ -63,7 +63,7 @@ void RGR_List::set_rgrs(int tar)
     {
         if (tar & 1)
         {
-			int rgr_tmp = pow(2.0, pos);
+			int rgr_tmp = static_cast<int>(pow(2.0, pos));
             this->rgr_list.push_back(RGR_Enum(rgr_tmp));
         }
